Add authentication filter

+++ auth-service/src/main/java/com/example/authservice/repository/UserRepository.java
@@ -1,10 +1,10 @@
 package com.example.authservice.repository;
 
-import com.example.authservice.model.User;
+import com.example.authservice.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 }
